Revert "containers: add practice container"
This reverts commit 467a6ee981afa33fefb41c42fde4fb2874f4b07f.
