+++ src/components/Title/style.module.css
@@ -55,7 +55,3 @@
 .medium-1 {
   font-size: 20px;
 }
-
-.mediumBold {
-  font-weight: 500;
-}
