--- src/containers/Practice/style.module.css
@@ -1,56 +1,0 @@
-@import '../../assets/css/colors.css';
-
-.container {
-  padding: 20px;
-  background-color: var(--color-white);
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  min-height: 80vh;
-}
-
-.header {
-  align-items: center;
-  display: flex;
-  padding: 20px 0;
-}
-
-.header > :first-child{
-  margin: 0;
-  min-height: 40px;
-  width: calc((100vh - 40px) / 2);
-}
-
-.header > :last-child{
-  min-height: 40px;
-}
-
-.headerClose {
-  width: 100%;
-}
-
-.content {
-  padding-top: 20px;
-  min-height: 320px;
-}
-
-.image {
-  height: 200px;
-  min-width: calc(310px - 40px);
-  overflow: hidden;
-  padding-bottom: 25px;
-  text-align: center;
-}
-
-.image img {
-  width: auto;
-  height: 100%;
-}
-
-.footer {
-  min-height: 120px;
-}
-
-.footer > * {
-  margin-top: 15px;
-}
