+++ src/containers/Practices/Practices.test.js
@@ -1,11 +1,12 @@
 import React from 'react'
 import { render } from '@testing-library/react'
 
-import PracticesContainer from '.'
+import PracticesContainer from './'
 
 const goodPractices = [
   {
-    title: "Resumo dos Procedimentos para Auditoria TCP",
+    title: "Guia de boas práticas",
+    subtitle: "Resumo dos Procedimentos para Auditoria TCP",
     message: "Material Básico: 5 Copos, 1 Par de luvas, Touca, Jarra, Pincel, Papel toalha, Balança e  Tabela de programação.",
     image: 'PT0',
   },
@@ -16,7 +17,7 @@ const goodPractices = [
   },
 ]
 
-describe('Practice container', () => {
+describe('Practices container', () => {
   it('should render correctly', () => {
     const callback = jest.fn()
     const step = 0
