+++ src/containers/Practices/index.js
@@ -1,71 +1,104 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-
-import Translate from '../../locales'
 import {
-  PracticeCard,
+  Button,
+  Progress,
   Title,
 } from '../../components'
+import Translate from '../../locales'
 import styles from './style.module.css'
+import ImagesPractice from './images'
+import CloseSvg from '../../assets/icons/close.svg'
 
-import ArrowSvg from '../../assets/icons/arrow-left-white.svg'
-import BannerSvg from '../../assets/images/banners/banner-2.svg'
-
-const Practices = ({
-  goBack,
-  practices,
-  goToPractice
+const Practice = ({
+  close,
+  onStep,
+  onPrevStep,
+  practice,
+  progress,
+  step,
+  steps,
 }) => {
-  const PracticeItem = ({
-    id,
-    image,
+  const {
     title,
-    readTime,
-  }) => (
-    <PracticeCard
-      action={() => goToPractice(id)}
-      key={id}
-      image={image}
-      title={title}
-      readTime={readTime}
-    />
+    message,
+    image,
+  } = practice
+
+  const ButtonText = (
+    step === 0
+      ? Translate('practices.start')
+      : Translate('practices.next')
   )
 
   return (
     <div className={styles.container}>
-      <div className={styles.header}>
-        <div
-          onClick={goBack}
-          role="button"
-        >
-          <img src={ArrowSvg} alt="go back" />
-        </div>
-        <img src={BannerSvg} alt="ilustration" />
+      <div
+        onClick={close}
+        role="button"
+        className={styles.headerClose}
+      >
+        <img src={CloseSvg} alt="close practices" />
       </div>
-      <div className={styles.content}>
+      <div className={styles.header}>
         <Title
           type="bold"
-          text={Translate('practices.title')}
+          text={title}
+          textAlign="center"
         />
-        <div className={styles.pratices}>
-          {practices.map(PracticeItem)}
+        <div className={styles.progressPosition}>
+          <Progress
+            progress={progress}
+            setup={{
+              type: 'radial',
+              radial: {
+                cx: 8,
+                cy: 8,
+                r: 8,
+              },
+            }}
+          />
         </div>
+      </div>
+      <div className={styles.content}>
+        <div className={styles.image}>
+          <img src={ImagesPractice[image]} alt="good practices" />
+        </div>
+        <Title
+          color="secondary"
+          text={message}
+          textAlign="center"
+          size="normal"
+        />
+      </div>
+      <div className={styles.footer}>
+        <Button color="primary" action={onStep}>
+          {step === steps ? Translate('practices.goToHome') : ButtonText}
+        </Button>
+        {
+          step > 0 && (
+            <Button color="outline" action={onPrevStep}>
+              {Translate('practices.previous')}
+            </Button>
+          )
+        }
       </div>
     </div>
   )
 }
 
-Practices.propTypes = {
-  goBack: PropTypes.func.isRequired,
-  goToPractice: PropTypes.func.isRequired,
-  practices: PropTypes.arrayOf(
-    PropTypes.shape({
-      id: PropTypes.number.isRequired,
-      image: PropTypes.string.isRequired,
-      title: PropTypes.string.isRequired,
-      readTime: PropTypes.string,
-    }).isRequired,
-  ).isRequired,
+Practice.propTypes = {
+  close: PropTypes.func.isRequired,
+  onStep: PropTypes.func.isRequired,
+  onPrevStep: PropTypes.func.isRequired,
+  progress: PropTypes.number.isRequired,
+  practice: PropTypes.shape({
+    title: PropTypes.string.isRequired,
+    message: PropTypes.string.isRequired,
+    image: PropTypes.string.isRequired,
+  }).isRequired,
+  step: PropTypes.number.isRequired,
+  steps: PropTypes.number.isRequired,
 }
 
-export default Practices
+export default Practice
