+++ src/utils/api/practices.mock.js
@@ -1,237 +1,229 @@
 const goodPractices = [
   {
-    id: 1,
-    image: 'BannerSvg3',
     title: "Resumo dos Procedimentos para Auditoria TCP",
-    readTime: '5-6 minutos de leitura',
-    practices: [
-      {
-        title: "Resumo dos Procedimentos para Auditoria TCP",
-        message: "Material Básico: 5 Copos, 1 Par de luvas, Touca, Jarra, Pincel, Papel toalha, Balança e  Tabela de programação.",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 1",
-        message: "Colocar a touca ao entrar no Ponto.",
-        image: 'PT1',
-      },
-      {
-        title: "Procedimento 2",
-        message: "Identificar equipamento visualmente, verificar a etiqueta de BMB e o PC (placa de património).",
-        image: 'PT2',
-      },
-      {
-        title: "Procedimento 3",
-        message: "Tirar testeira e tabela.",
-        image: 'PT3',
-      },
-      {
-        title: "Procedimento 4",
-        message: "Verificar display da porta se tem alguma mensagem.",
-        image: 'PT4',
-      },
-      {
-        title: "Procedimento 5",
-        message: "Abrir a porta.        ",
-        image: 'PT5',
-      },
-      {
-        title: "Procedimento 6",
-        message: "Levantar a tampa superior.        ",
-        image: 'PT6',
-      },
-      {
-        title: "Procedimento 7",
-        message: "Procurar sinais de fauna nociva.",
-        image: 'PT7',
-      },
-      {
-        title: "Procedimento 8",
-        message: "Verificar o nível dos produtos nos canisters (Caso seja necessário abastecer usar pacotes já abertos).",
-        image: 'PT8',
-      },
-      {
-        title: "Procedimento 9",
-        message: "Colocar Luvas.",
-        image: 'PT9',
-      },
-      {
-        title: "Procedimento 10",
-        message: " Fechar as travas das calhas de todos os canisters.        ",
-        image: 'PT10',
-      },
-      {
-        title: "Procedimento 11",
-        message: " Conferir a limpeza da calha (retirar resíduos ou lavar).",
-        image: 'PT11',
-      },
-      {
-        title: "Procedimento 12",
-        message: " Retirar os canisters, primeiro o maior e depois os outros, virá-los e dar leves batidas na parte interior do mesmo.",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 13",
-        message: " Colocar os canisters menores apoiados no equipamento e depois o maior para evitar tombamento (Somente aplicável para solução solúvel). ",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 14",
-        message: " Observar indícios de abastecimento incorreto, caso ocorra reforçar treinamento com cliente.",
-        image: 'PT14',
-      },
-      {
-        title: "Procedimento 15",
-        message: " Abastecer os canisters fora do equipamento.        ",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 16",
-        message: " Limpar a plataforma (Verificar os motores de produto se necessário troca-los ou repara-los).",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 17",
-        message: " Executar a lavagem.",
-        image: 'PT17',
-      },
-      {
-        title: "Procedimento 18",
-        message: " Retirar os copos misturadores e fazer a limpeza.",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 19",
-        message: " Retirar a bandeja do exaustor para limpeza e recoloca-la novamente no lugar.",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 20",
-        message: " Recolocar os canisters.",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 21",
-        message: " Retirar as hélices com chave de fenda e verificar e trocar oringues e guarnições se necessário.",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 22",
-        message: " Limpar parte de trás das hélices e a frente do motor.",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 23",
-        message: " Montar as hélices, guarnições e orings.",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 24",
-        message: " Colocar os copos misturadores.",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 25",
-        message: " Separar 06 copos. ",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 26",
-        message: " Posicionar os copos (3 para os produtos e 2 para a água).",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 27",
-        message: " Inverter a Posição das calhas (Somente aplicável para solução solúvel). ",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 28",
-        message: " Todo o produto que cair no copo no momento da abertura das travas deverá ser devolvido no canister correspondente,caso misture dois produtos no copo joga-lo fora.",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 29",
-        message: " Proceda a gramatura.",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 30",
-        message: " Terminada a gramatura sair do sistema.",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 31",
-        message: " Remover, lavar, secar e recolocar as calhas em suas posições. ",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 32",
-        message: " Abrir as travas das calhas.",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 33",
-        message: " Fazer a lavagem.",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 34",
-        message: " Limpar o suporte das mangueiras com pincel ou lavar (trocar as mangueiras se necessário).",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 35",
-        message: " Retirar a bandeja de resíduo e lavar.",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 36",
-        message: " Limpar a base inferior com pincel ou papel toalha.",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 37",
-        message: " Abaixar a tampa.",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 38",
-        message: " Fechar a porta.",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 39",
-        message: " Verificar limpeza externa do equipamento.",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 40",
-        message: " Verificar mangueiras sujas e trocar se necessário.",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 41",
-        message: " Colocar a testeira e a tabela.",
-        image: 'PT0',
-      },
-      {
-        title: "Procedimento 42",
-        message: " Dar prosseguimento à auditoria (BMB).",
-        image: 'PT42',
-      },
-      {
-        title: "Procedimento 43",
-        message: " Tirar dose no modo teste.",
-        image: 'PT43',
-      },
-      {
-        title: "Procedimento 44",
-        message: " Visitar o estoque do cliente.",
-        image: 'PT44',
-      },
-    ]
-  }
+    message: "Material Básico: 5 Copos, 1 Par de luvas, Touca, Jarra, Pincel, Papel toalha, Balança e  Tabela de programação.",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 1",
+    message: "Colocar a touca ao entrar no Ponto.",
+    image: 'PT1',
+  },
+  {
+    title: "Procedimento 2",
+    message: "Identificar equipamento visualmente, verificar a etiqueta de BMB e o PC (placa de património).",
+    image: 'PT2',
+  },
+  {
+    title: "Procedimento 3",
+    message: "Tirar testeira e tabela.",
+    image: 'PT3',
+  },
+  {
+    title: "Procedimento 4",
+    message: "Verificar display da porta se tem alguma mensagem.",
+    image: 'PT4',
+  },
+  {
+    title: "Procedimento 5",
+    message: "Abrir a porta.        ",
+    image: 'PT5',
+  },
+  {
+    title: "Procedimento 6",
+    message: "Levantar a tampa superior.        ",
+    image: 'PT6',
+  },
+  {
+    title: "Procedimento 7",
+    message: "Procurar sinais de fauna nociva.",
+    image: 'PT7',
+  },
+  {
+    title: "Procedimento 8",
+    message: "Verificar o nível dos produtos nos canisters (Caso seja necessário abastecer usar pacotes já abertos).",
+    image: 'PT8',
+  },
+  {
+    title: "Procedimento 9",
+    message: "Colocar Luvas.",
+    image: 'PT9',
+  },
+  {
+    title: "Procedimento 10",
+    message: " Fechar as travas das calhas de todos os canisters.        ",
+    image: 'PT10',
+  },
+  {
+    title: "Procedimento 11",
+    message: " Conferir a limpeza da calha (retirar resíduos ou lavar).",
+    image: 'PT11',
+  },
+  {
+    title: "Procedimento 12",
+    message: " Retirar os canisters, primeiro o maior e depois os outros, virá-los e dar leves batidas na parte interior do mesmo.",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 13",
+    message: " Colocar os canisters menores apoiados no equipamento e depois o maior para evitar tombamento (Somente aplicável para solução solúvel). ",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 14",
+    message: " Observar indícios de abastecimento incorreto, caso ocorra reforçar treinamento com cliente.",
+    image: 'PT14',
+  },
+  {
+    title: "Procedimento 15",
+    message: " Abastecer os canisters fora do equipamento.        ",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 16",
+    message: " Limpar a plataforma (Verificar os motores de produto se necessário troca-los ou repara-los).",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 17",
+    message: " Executar a lavagem.",
+    image: 'PT17',
+  },
+  {
+    title: "Procedimento 18",
+    message: " Retirar os copos misturadores e fazer a limpeza.",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 19",
+    message: " Retirar a bandeja do exaustor para limpeza e recoloca-la novamente no lugar.",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 20",
+    message: " Recolocar os canisters.",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 21",
+    message: " Retirar as hélices com chave de fenda e verificar e trocar oringues e guarnições se necessário.",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 22",
+    message: " Limpar parte de trás das hélices e a frente do motor.",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 23",
+    message: " Montar as hélices, guarnições e orings.",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 24",
+    message: " Colocar os copos misturadores.",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 25",
+    message: " Separar 06 copos. ",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 26",
+    message: " Posicionar os copos (3 para os produtos e 2 para a água).",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 27",
+    message: " Inverter a Posição das calhas (Somente aplicável para solução solúvel). ",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 28",
+    message: " Todo o produto que cair no copo no momento da abertura das travas deverá ser devolvido no canister correspondente,caso misture dois produtos no copo joga-lo fora.",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 29",
+    message: " Proceda a gramatura.",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 30",
+    message: " Terminada a gramatura sair do sistema.",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 31",
+    message: " Remover, lavar, secar e recolocar as calhas em suas posições. ",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 32",
+    message: " Abrir as travas das calhas.",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 33",
+    message: " Fazer a lavagem.",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 34",
+    message: " Limpar o suporte das mangueiras com pincel ou lavar (trocar as mangueiras se necessário).",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 35",
+    message: " Retirar a bandeja de resíduo e lavar.",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 36",
+    message: " Limpar a base inferior com pincel ou papel toalha.",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 37",
+    message: " Abaixar a tampa.",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 38",
+    message: " Fechar a porta.",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 39",
+    message: " Verificar limpeza externa do equipamento.",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 40",
+    message: " Verificar mangueiras sujas e trocar se necessário.",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 41",
+    message: " Colocar a testeira e a tabela.",
+    image: 'PT0',
+  },
+  {
+    title: "Procedimento 42",
+    message: " Dar prosseguimento à auditoria (BMB).",
+    image: 'PT42',
+  },
+  {
+    title: "Procedimento 43",
+    message: " Tirar dose no modo teste.",
+    image: 'PT43',
+  },
+  {
+    title: "Procedimento 44",
+    message: " Visitar o estoque do cliente.",
+    image: 'PT44',
+  },
 ]
 
 export default goodPractices
